fix: moving blocks to elements
