+++ lib/widgets/qu-reviews/src/edit.js
@@ -1,7 +1,8 @@
 import { BlockControls, useBlockProps } from '@wordpress/block-editor';
 
-import LayoutPicker from './components/LayoutPicker';
-import ReviewSelect from './components/ReviewSelect';
+import LayoutPicker from './elements/LayoutPicker';
+import ReviewSelect from './elements/ReviewSelect';
+import MaxReviews from './elements/MaxReviews';
 import ReviewsSorted from './layouts/ReviewsSorted';
 
 import './scss/editor.scss';
@@ -9,7 +10,7 @@ import './scss/editor.scss';
 const { registerBlockType } = wp.blocks;
 const { withSelect } = wp.data;
 
-const { TextControl, ToolbarGroup } = wp.components;
+const { Toolbar, ToolbarGroup } = wp.components;
 
 registerBlockType( 'qu/reviews', {
 	apiVersion: 2,
@@ -42,7 +43,7 @@ registerBlockType( 'qu/reviews', {
 	} )( ( props ) => {
 		const blockProps = useBlockProps();   //eslint-disable-line
 		const { categories, attributes, setAttributes } = props;
-		const { categoryId, maxReviews } = attributes;
+		const { categoryId } = attributes;
 
 		if ( ! categories ) {
 			return (
@@ -67,29 +68,24 @@ registerBlockType( 'qu/reviews', {
 		return (
 			<div { ...blockProps }>
 				<BlockControls>
-					<ToolbarGroup label="toolbar">
-						{ categoryId
-							? <ReviewSelect { ...props } />
-							: null
-						}
-						{
-							categoryId
-								? <LayoutPicker { ...props } />
+					<Toolbar label="toolbar" className="qu-Reviews__edit">
+						<ToolbarGroup>
+							{ categoryId
+								? <ReviewSelect { ...props } />
 								: null
-						}
-						{
-							categoryId
-								? <TextControl
-										label="Show max.:"
-										type="number"
-										min="1"
-										max="20"
-										value={ maxReviews }
-										onChange={ ( value ) => setAttributes( { maxReviews: value } ) }
-								/>
-								: null
-						}
-					</ToolbarGroup>
+							}
+							{
+								categoryId
+									? <LayoutPicker { ...props } />
+									: null
+							}
+							{
+								categoryId
+									? <MaxReviews { ...props } />
+									: null
+							}
+						</ToolbarGroup>
+					</Toolbar>
 				</BlockControls>
 				{ categoryId
 					? <ReviewsSorted { ...props } />
