+++ lib/widgets/qu-reviews/src/elements/LayoutPicker.js
